integrated bounding boxes for meshes
also removed mention of dependency folder from includes

+++ manager/data/deserialization.cpp
@@ -33,7 +33,7 @@ string ManagerData::Deserialization::TEXT(string Field)
 
 int ManagerData::Deserialization::Index(string Field)
 {
-	for(unsigned int i = 0; i < fields.size(); ++i)
+	for(size_t i = 0; i < fields.size(); ++i)
 		if(fields[i] == Field)
 			return i;
 	return -1;
